build: tighten ruff ruleset (B, SIM, RUF, PIE, ISC)
Enable flake8-bugbear, flake8-simplify, ruff-specific, flake8-pie and
implicit-string-concat rule groups, and resolve the resulting findings.

Policy ignores (documented in pyproject.toml): SIM108, SIM105, RUF059,
RUF002, ISC001 (formatter conflict), and SIM115 (open() without a context
manager - real but needs careful per-site I/O refactoring; tracked as a
follow-up).

Fixes applied:
- Simplifications: return conditions directly, 'a != b', merged isinstance,
  'key in dict', tuple startswith, iterable unpacking, enumerate() loop.
- Bugbear: 'raise ... from excep'; replace a useless 'self.seen_binding'
  expression (which caught the wrong exception type) with an explicit
  hasattr() initialisation; rename unused loop variables.
- Drop redundant int(round(...)) casts (round() already returns int).

No behavioural changes intended.

+++ superpaper/perspective.py
@@ -75,7 +75,7 @@ def get_backprojected_display_system(crops, persp_data, plot=False):
     for quad in projected_quads:
         work_quad = []
         for corner in quad:
-            work_quad.append((int(round(corner[0] - leftmost_corner)), int(round(corner[1] - bottommost_corner))))
+            work_quad.append((round(corner[0] - leftmost_corner), round(corner[1] - bottommost_corner)))
         translated_quads.append(work_quad)
     sp_logging.G_LOGGER.info("translated_quads: %s", translated_quads)
     ordered_quads = [(tquad[2], tquad[3], tquad[1], tquad[0]) for tquad in translated_quads]
